Advanced Route protection and user personalisation after auth

+++ backend/controllers/mainController.js
@@ -1,11 +1,13 @@
 import asyncHandler from 'express-async-handler';
 import Goal from '../models/goalModel.js';
+import User from '../models/userModel.js';
+
 
 //@desc   GET goals
 //@route  api/goals
 //access  PRIVATE
 export const getGoals = asyncHandler(async (req,res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({user: req.user.id});
 
     res.status(200).json(goals);
 })
@@ -29,6 +31,7 @@ export const setGoal = async (req,res,next) => {
     }
 
     const goal = await Goal.create({
+        user: req.user.id,
         text: req.body.goal
     })
     res.status(201).json(goal);
@@ -44,11 +47,20 @@ export const updateGoal = async (req,res) => {
 
     const goal = await Goal.findById(id);
 
+    const user = await User.findById(req.user.id);
+
     if(!goal)
     {
-        res.status(400);
-        throw new Error('Goal Not Set');
+        res.status(400)
+        throw new Error("Goal Not Set");
     }
+
+    if(goal.user.toString() != user.id)
+    {
+        res.status(401)
+        throw new Error("You cannot update a goal that is not yours");
+    }
+
 
     const updatedGoal = await Goal.findByIdAndUpdate(id, { text: req.body.goal}, {
         new: true
@@ -69,6 +81,14 @@ export const deleteGoal = async (req,res) => {
     {
         res.status(400);
         throw new Error(`Goal with id: ${id} not found`);
+    }
+
+    const user = await User.findById(req.user.id);
+
+    if(goal.user.toString() != user.id)
+    {
+        res.status(401)
+        throw new Error("You cannot delete a goal that is not yours");
     }
 
     await goal.deleteOne();
